(refactor) structure for new plugin sys
